migrated location setting to fixed station section
