+++ app/src/main/java/hpsaturn/pollutionreporter/view/SettingsFixedStation.java
@@ -2,29 +2,28 @@ package hpsaturn.pollutionreporter.view;
 
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.os.Handler;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.preference.ListPreference;
-import androidx.preference.Preference;
 import androidx.preference.SwitchPreference;
 
 import com.hpsaturn.tools.Logger;
 
+import java.text.DecimalFormat;
+
 import hpsaturn.pollutionreporter.R;
-import hpsaturn.pollutionreporter.models.ApiConfig;
 import hpsaturn.pollutionreporter.models.CommandConfig;
+import hpsaturn.pollutionreporter.models.GeoConfig;
 import hpsaturn.pollutionreporter.models.InfluxdbConfig;
 import hpsaturn.pollutionreporter.models.ResponseConfig;
-import hpsaturn.pollutionreporter.models.SampleConfig;
 import hpsaturn.pollutionreporter.models.SensorConfig;
-import hpsaturn.pollutionreporter.models.SensorData;
 import hpsaturn.pollutionreporter.models.SensorName;
-import hpsaturn.pollutionreporter.models.SensorType;
 import hpsaturn.pollutionreporter.models.WifiConfig;
+import io.nlopez.smartlocation.SmartLocation;
 
 /**
  * Created by Antonio Vanegas @hpsaturn on 2/17/19.
@@ -45,12 +44,14 @@ public class SettingsFixedStation extends SettingsBaseFragment {
     }
 
     public void refreshUI(){
+        Logger.i(TAG,"[Config] refreshUI");
         updateSensorNameSummary();
         updateWifiSummary();
-        updateApiHostSummary();
-        updateApiUriSummary();
         updateInfluxDbSummmary();
         updateInfluxPortSummary();
+        updateLocationSummary();
+        lastLocation = SmartLocation.with(getActivity()).location().getLastLocation();
+        validateLocationSwitch();
     }
 
     @Override
@@ -73,12 +74,10 @@ public class SettingsFixedStation extends SettingsBaseFragment {
         if (config.ienb != getInfluxDbSwitch().isChecked()) {
             notify_sync = true;
         }
-        if (config.aenb != getApiSwitch().isChecked()) {
-            notify_sync = true;
-        }
 
-        saveDeviceInfoString(config);
         setStatusSwitch(true);
+        updateLocationSummary();
+        validateLocationSwitch();
 
         if (notify_sync) {
             saveAllPreferences(config);
@@ -99,7 +98,6 @@ public class SettingsFixedStation extends SettingsBaseFragment {
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        infoPreferenceInit();
     }
 
     @Override
@@ -115,22 +113,14 @@ public class SettingsFixedStation extends SettingsBaseFragment {
                 getWifiSwitch().setEnabled(isWifiSwitchFieldsValid());
             } else if (key.equals(getString(R.string.key_setting_enable_wifi))) {
                 saveWifiConfig();
-            } else if (key.equals(getString(R.string.key_setting_apiusr))) {
-                getApiSwitch().setEnabled(isApiSwitchFieldsValid());
-            } else if (key.equals(getString(R.string.key_setting_apipss))) {
-                getApiSwitch().setEnabled(isApiSwitchFieldsValid());
-            } else if (key.equals(getString(R.string.key_setting_apisrv))) {
-                getApiSwitch().setEnabled(isApiSwitchFieldsValid());
-            } else if (key.equals(getString(R.string.key_setting_enable_api))) {
-                saveApiConfig();
             } else if (key.equals(getString(R.string.key_setting_ifxdb))) {
                 getInfluxDbSwitch().setEnabled(isInfluxDbSwitchFieldsValid());
             } else if (key.equals(getString(R.string.key_setting_ifxip))) {
                 getInfluxDbSwitch().setEnabled(isInfluxDbSwitchFieldsValid());
             } else if (key.equals(getString(R.string.key_setting_enable_ifx))) {
                 saveInfluxDbConfig();
-            } else if (key.equals(getString(R.string.key_setting_vars))) {
-                getMain().selectedVarsUpdated();
+            } else if (key.equals(getString(R.string.key_setting_enable_location))) {
+                saveLocation();
             }
 
         }
@@ -216,62 +206,6 @@ public class SettingsFixedStation extends SettingsBaseFragment {
         sendSensorConfig(config);
     }
 
-    /***********************************************************************************************
-     * API switch
-     **********************************************************************************************/
-
-    private void saveApiConfig() {
-
-        if (getApiSwitch().isChecked() && isApiSwitchFieldsValid()) {
-            ApiConfig config = new ApiConfig();
-            config.apiusr = getSharedPreference(getString(R.string.key_setting_apiusr));;
-            config.apipss = getSharedPreference(getString(R.string.key_setting_apipss));
-            config.apisrv = getSharedPreference(getString(R.string.key_setting_apisrv));
-            config.apiuri = getSharedPreference(getString(R.string.key_setting_apiuri));
-            config.apiprt = 80;  // TODO: sending via UI
-            config.aenb = true;
-            sendSensorConfig(config);
-        } else
-            setApiSwitch(false);
-    }
-
-    private boolean isApiSwitchFieldsValid(){
-        Logger.v(TAG, "[Config] validating ->" + getString(R.string.key_setting_enable_api));
-        String api_usr = getSharedPreference(getString(R.string.key_setting_apiusr));
-        String api_pss = getSharedPreference(getString(R.string.key_setting_apipss));
-        String api_srv = getSharedPreference(getString(R.string.key_setting_apisrv));
-        Logger.v(TAG, "[Config] values -> " + api_usr +  " " + api_pss + " " + api_srv);
-        apiSummaryUpdate();
-        return !(api_usr.length() == 0 || api_pss.length() == 0 || api_srv.length() == 0);
-    }
-
-    private void setApiSwitch(boolean checked) {
-        Logger.v(TAG, "[Config] API switch check -> " + checked);
-        SwitchPreference apiSwitch = getApiSwitch();
-        apiSwitch.setEnabled(isApiSwitchFieldsValid());
-        apiSwitch.setChecked(checked);
-        apiSummaryUpdate();
-        enableApiOnDevice(checked);
-    }
-
-    private void apiSummaryUpdate(){
-        updateApiUsrSummmary();
-        updateApiUriSummary();
-        updateApiHostSummary();
-        updatePasswSummary(R.string.key_setting_apipss);
-    }
-
-    private SwitchPreference getApiSwitch() {
-        return findPreference(getString(R.string.key_setting_enable_api));
-    }
-
-    private void enableApiOnDevice(boolean enable) {
-        CommandConfig config = new CommandConfig();
-        config.cmd = getSharedPreference(getString(R.string.key_setting_wmac));
-        config.act = "ast";
-        config.aenb = enable;
-        sendSensorConfig(config);
-    }
 
     /***********************************************************************************************
      * InfluxDB switch
@@ -329,44 +263,50 @@ public class SettingsFixedStation extends SettingsBaseFragment {
     }
 
     /***********************************************************************************************
-     * Device info
-     **********************************************************************************************/
-
-    private void infoPreferenceInit() {
-        Preference infoPreference = findPreference(getString(R.string.key_device_info));
-
-        infoPreference.setOnPreferenceClickListener(preference -> {
-            readSensorConfig();
-            return false;
-        });
-    }
-
-    public void onReadSensorData(SensorData data) {
-        String info = getSharedPreference(getString(R.string.key_device_info));
-        info = info + "\n-----------------------"
-                +"\n"+data.dsl+"->PM2.5:"+data.P25+" PM1:"+data.P1+" CO2:"+data.CO2
-                + "\n T:"+data.tmp+" H:"+data.hum;
-
-        updateSummary(R.string.key_device_info,info);
-        saveSharedPreference(R.string.key_device_info,info);
-    }
-
-    private void saveDeviceInfoString(ResponseConfig config) {
-        String info = "MAC:"
-                + config.vmac
-                +"\nFirmware: "+config.vflv+" rev"+config.vrev+" ("+config.vtag+")"
-                +"\n[WiFi:"+(config.wenb ? "On" : "Off")+"][IFDB:"+(config.ienb ? "On" : "Off")
-                +"][GW:"+(config.wsta ? "On" : "Off")+"]";
-        if(config.vrev<774)info="\n!!YOUR FIRMWARE IS OUTDATED!!\n\n"+info;
-        updateSummary(R.string.key_device_info,info);
-        saveSharedPreference(R.string.key_device_info,info);
-    }
-
-
-
-    /***********************************************************************************************
      * Misc preferences section
      **********************************************************************************************/
+
+    private void validateLocationSwitch() {
+        SwitchPreference locationSwitch = findPreference(getString(R.string.key_setting_enable_location));
+        locationSwitch.setEnabled(lastLocation!=null);
+    }
+
+    private void saveLocation() {
+        SwitchPreference locationSwitch = findPreference(getString(R.string.key_setting_enable_location));
+        if(lastLocation != null) {
+            if(locationSwitch.isChecked()) {
+                snackBar = getMain().getSnackBar(R.string.msg_set_current_location, R.string.bt_location_save_action, view -> {
+                    getMain().showSnackMessage(R.string.msg_save_location);
+                    GeoConfig config = new GeoConfig();
+                    config.lat = lastLocation.getLatitude();
+                    config.lon = lastLocation.getLongitude();
+                    config.alt = lastLocation.getAltitude();
+                    config.spd = lastLocation.getSpeed();
+                    sendSensorConfig(config);
+                    Handler handler = new Handler();
+                    handler.postDelayed(() -> locationSwitch.setChecked(false), 2000);
+                });
+                snackBar.show();
+            }
+            else{
+                snackBar.dismiss();
+            }
+        }
+        else {
+            getMain().showSnackMessage(R.string.msg_save_location_failed);
+        }
+        updateLocationSummary();
+    }
+
+    private void updateLocationSummary() {
+        if (lastLocation != null) {
+            DecimalFormat precision = new DecimalFormat("0.000");
+            String accu = "Accu:" + (int) lastLocation.getAccuracy() + "m ";
+            String lat = "(" + precision.format(lastLocation.getLatitude());
+            String lon = "," + precision.format(lastLocation.getLongitude()) + ")";
+            updateSummary(R.string.key_setting_enable_location,accu + lat + lon);
+        }
+    }
 
     /***********************************************************************************************
      * Update methods
@@ -375,14 +315,10 @@ public class SettingsFixedStation extends SettingsBaseFragment {
 
     private void updatePreferencesSummmary(ResponseConfig config) {
         if(config.dname !=null)updateSummary(R.string.key_setting_dname,config.dname);
-        if(config.apiusr !=null)updateSummary(R.string.key_setting_apiusr,config.apiusr);
-        if(config.apisrv !=null)updateSummary(R.string.key_setting_apisrv,config.apisrv);
-        if(config.apiuri !=null)updateSummary(R.string.key_setting_apiuri,config.apiuri);
         if(config.ssid !=null)updateSummary(R.string.key_setting_ssid,config.ssid);
         if(config.ifxdb !=null)updateSummary(R.string.key_setting_ifxdb,config.ifxdb);
         if(config.ifxip !=null)updateSummary(R.string.key_setting_ifxip,config.ifxip);
         updatePasswSummary(R.string.key_setting_pass);
-        updatePasswSummary(R.string.key_setting_apipss);
 
     }
 
@@ -400,18 +336,6 @@ public class SettingsFixedStation extends SettingsBaseFragment {
         updateSummary(R.string.key_setting_ifxip);
     }
 
-    private void updateApiUsrSummmary() {
-        updateSummary(R.string.key_setting_apiusr);
-    }
-
-    private void updateApiHostSummary() {
-        updateSummary(R.string.key_setting_apisrv);
-    }
-
-    private void updateApiUriSummary() {
-        updateSummary(R.string.key_setting_apiuri);
-    }
-
     private void updateSwitch(int key,boolean enable){
         SwitchPreference mSwitch = findPreference(getString(key));
         mSwitch.setChecked(enable);
@@ -420,7 +344,6 @@ public class SettingsFixedStation extends SettingsBaseFragment {
     private void updateSwitches(SensorConfig config){
         updateSwitch(R.string.key_setting_enable_wifi,config.wenb);
         updateSwitch(R.string.key_setting_enable_ifx,config.ienb);
-        updateSwitch(R.string.key_setting_enable_api,config.aenb);
     }
 
     /***********************************************************************************************
@@ -432,12 +355,7 @@ public class SettingsFixedStation extends SettingsBaseFragment {
         saveSharedPreference(R.string.key_setting_ssid, config.ssid);
         saveSharedPreference(R.string.key_setting_ifxdb, config.ifxdb);
         saveSharedPreference(R.string.key_setting_ifxip, config.ifxip);
-        saveSharedPreference(R.string.key_setting_apiusr, config.apiusr);
-        saveSharedPreference(R.string.key_setting_apisrv, config.apisrv);
         saveSharedPreference(R.string.key_setting_wmac, "" + config.wmac);
-        saveDeviceInfoString(config);
     }
-
-
 
 }
