+++ app/src/main/res/xml/settings_device_tools.xml
@@ -2,22 +2,20 @@
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <PreferenceCategory app:title="Device Status">
+    <PreferenceCategory>
         <SwitchPreference
+            app:defaultValue="false"
             app:icon="@android:drawable/ic_lock_power_off"
             app:key="@string/key_device_status"
             app:summary="Disconnected"
-            app:defaultValue="false"
             app:title="Device Status:" />
     </PreferenceCategory>
 
 
-    <PreferenceCategory app:title="Some Device Tools"
-        android:dependency="@string/key_device_status">
+    <PreferenceCategory
+        android:dependency="@string/key_device_status"
+        app:title="Some Device Tools">
 
-        <SwitchPreference
-            app:key="@string/key_setting_enable_location"
-            app:title="Save Current Location" />
 
         <SwitchPreference
             app:defaultValue="false"
