+++ app/src/main/res/xml/settings_fixed_station.xml
@@ -2,7 +2,7 @@
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <PreferenceCategory app:title="Device Status">
+    <PreferenceCategory>
         <SwitchPreference
             app:defaultValue="false"
             app:icon="@android:drawable/ic_lock_power_off"
@@ -46,7 +46,7 @@
 
     <PreferenceCategory
         app:dependency="@string/key_setting_enable_wifi"
-        app:title="InfluxDB Cloud">
+        app:title="CanAirIO influxDB">
 
         <EditTextPreference
             android:inputType="text"
@@ -69,66 +69,17 @@
 
         <SwitchPreference
             app:key="@string/key_setting_enable_ifx"
-            app:title="Enable InfluxDB Cloud" />
+            app:summary="Publish to CanAirIO or custom server"
+            app:title="Enable InfluxDB" />
 
     </PreferenceCategory>
 
+
     <PreferenceCategory
-        app:dependency="@string/key_setting_enable_wifi"
-        app:isPreferenceVisible="false"
-        app:title="Deprecated CanAirIO API">
-
-        <EditTextPreference
-            android:inputType="text"
-            app:icon="@android:drawable/ic_menu_upload"
-            app:key="@string/key_setting_apiusr"
-            app:summary=""
-            app:title="Username" />
-
-        <EditTextPreference
-            android:inputType="textPassword"
-            app:key="@string/key_setting_apipss"
-            app:summary="CanAirIO API Password"
-            app:title="Password" />
-
-        <EditTextPreference
-            android:defaultValue="@string/value_apisrv_default"
-            android:inputType="text"
-            app:key="@string/key_setting_apisrv"
-            app:summary=""
-            app:title="Hostname" />
-
-        <EditTextPreference
-            android:defaultValue="@string/value_apiuri_default"
-            android:inputType="text"
-            app:key="@string/key_setting_apiuri"
-            app:summary=""
-            app:title="API target" />
+        app:title="Save current location">
 
         <SwitchPreference
-            app:key="@string/key_setting_enable_api"
-            app:title="Enable CanAirIO API" />
-
+            app:key="@string/key_setting_enable_location"
+            app:title="Send Lat/Long to device" />
     </PreferenceCategory>
-
-
-    <PreferenceCategory
-        app:isPreferenceVisible="false"
-        app:title="">
-        <Preference
-            app:icon="@android:drawable/ic_menu_send"
-            app:key="@string/key_send_feedback"
-            app:summary="Report technical issues or suggest new features"
-            app:title="Send Feedback" />
-    </PreferenceCategory>
-
-    <PreferenceCategory app:title="">
-        <Preference
-            app:icon="@android:drawable/ic_popup_sync"
-            app:key="@string/key_device_info"
-            app:summary=""
-            app:title="Device info:" />
-    </PreferenceCategory>
-
-
 </PreferenceScreen>
