+++ app/src/main/res/xml/settings_general.xml
@@ -67,4 +67,21 @@
         </Preference>
     </PreferenceCategory>
 
+    <PreferenceCategory app:title="">
+        <Preference
+            app:icon="@android:drawable/ic_popup_sync"
+            app:key="@string/key_device_info"
+            app:summary=""
+            app:title="Device info:" />
+    </PreferenceCategory>
+
+    <PreferenceCategory
+        app:title="">
+        <Preference
+            app:icon="@android:drawable/ic_menu_send"
+            app:key="@string/key_send_feedback"
+            app:summary="Report technical issues or suggest new features"
+            app:title="Send Feedback" />
+    </PreferenceCategory>
+
 </PreferenceScreen>
